Adding error message, feed implemented

+++ LITReviews/views.py
@@ -6,7 +6,7 @@ from django.db import IntegrityError
 from django.db.models import Count, CharField, Value
 from itertools import chain
 from django.shortcuts import get_object_or_404
-from django.http import HttpResponseNotAllowed, HttpResponse
+from django.contrib.auth.decorators import login_required
 
 
 def home_view(request):
@@ -14,14 +14,18 @@ def home_view(request):
 
 
 def subscription_view(request):
+    """Renders the subscription page and handles the follow and search operations."""
     error = False
     error_message = ""
     user = request.user
     followed_users = UserFollows.objects.filter(user=user.id)
     followers = UserFollows.objects.filter(followed_user=user.id)
+    users = User.objects.all()
     search_users = None
     if request.method == "POST":
-        search_query = request.POST.get("search")
+        search_query = request.POST.get("searchs")
+        print("debug")
+        print(search_query)
         if search_query:
             # Search operation
             search_users = (
@@ -49,26 +53,43 @@ def subscription_view(request):
                 error_message = (
                     f"Vous suivez déjà l'utilisateur {followed_user.username}"
                 )
+    context = {
+        "search_users": search_users,
+        "followers": followers,
+        "followed_users": followed_users,
+        "users":users,
+        "error": error,
+        "error_message": error_message,
+    }
 
-    return render(
-        request,
-        "LITReviews/subscription.html",
-        {
-            "search_users": search_users,
-            "followers": followers,
-            "followed_users": followed_users,
-            "error": error,
-            "error_message": error_message,
-        },
-    )
+    return render(request, "LITReviews/subscription.html", context)
+
 
 def unfollow_user(request):
+    """Handles the unfollow operation for a user"""
+    error = False
+    error_message = ""
+    user = request.user
+    followed_users = UserFollows.objects.filter(user=user.id)
+    followers = UserFollows.objects.filter(followed_user=user.id)
+    search_users = None
     if request.method == "POST":
-        user = request.user
         user_id = int(request.POST.get("user_id"))
-        UserFollows.objects.filter(user=user, followed_user_id=user_id).delete()
-        # Redirect back to the subscription page
-        return redirect("subscription")
+        followed_users = UserFollows.objects.filter(user=user.id)
+        if user_id in followed_users:
+            UserFollows.objects.filter(user=user, followed_user_id=user_id).delete()
+        else:
+            error = True
+            error_message = f"Vous ne suivez pas cet utilisateur"
+    context = {
+        "search_users": search_users,
+        "followers": followers,
+        "followed_users": followed_users,
+        "error": error,
+        "error_message": error_message,
+    }
+
+    return render(request, "LITReviews/subscription.html", context)
 
 
 def create_ticket(request):
@@ -86,21 +107,23 @@ def create_ticket(request):
 
 
 def ticket_detail(request, ticket_id):
+    error = False
+    error_message = ""
     ticket = Ticket.objects.get(id=ticket_id)
-    return render(request, "LITReviews/ticket_detail.html", {"ticket": ticket})
+    context = {
+        "ticket": ticket,
+        "error": error,
+        "error_message": error_message,
+    }
+    return render(request, "LITReviews/ticket_detail.html", context)
 
 
-def post_view(request):
-    # Get the current user
-    user = request.user
-
-    tickets = Ticket.objects.filter(user=user)
-
+def get_user_posts(users):
     # Get the tickets created by the user
-    user_tickets = Ticket.objects.filter(user=user)
+    user_tickets = Ticket.objects.filter(user__in=users)
     user_tickets = user_tickets.annotate(content_type=Value("TICKET", CharField()))
     # Get the reviews written by the user
-    user_reviews = Review.objects.filter(user=user)
+    user_reviews = Review.objects.filter(user__in=users)
     user_reviews = user_reviews.annotate(content_type=Value("REVIEW", CharField()))
 
     # Combine the tickets and reviews using the chain function
@@ -109,6 +132,37 @@ def post_view(request):
         key=lambda post: post.time_created,
         reverse=True,
     )
+
+    return user_posts
+
+
+def feed(request):
+    user = request.user
+
+    # Get the IDs of the users that the current user follows and the current user
+    feed_user_ids = list(
+        UserFollows.objects.filter(user=user).values_list("followed_user_id", flat=True)
+    )
+    feed_user_ids.append(user.id)
+
+    # Filter the tickets created by the user and his followers
+    tickets = Ticket.objects.filter(user__in=feed_user_ids)
+
+    user_posts = get_user_posts(feed_user_ids)
+
+    return render(
+        request,
+        "LITReviews/feed.html",
+        {"user_posts": user_posts, "tickets": tickets},
+    )
+
+
+def post_view(request):
+    # Get the current user
+    user = request.user
+    tickets = Ticket.objects.filter(user=user)
+
+    user_posts = user_posts = get_user_posts([user.id])
 
     return render(
         request,
@@ -136,20 +190,34 @@ def create_review(request, ticket_id):
     )
 
 
+@login_required
 def modify_ticket(request, ticket_id):
+    error = False
+    error_message = ""
     ticket = Ticket.objects.get(id=ticket_id)
+
+    # Check if the current user is the writer of the ticket or a superuser
+    if ticket.user != ticket.user and not request.user.is_superuser:
+        error = True
+        error_message = "Vous n'êtes pas autorisé(e) à modifier ce billet"
+        context = {
+            "ticket": ticket,
+            "error": error,
+            "error_message": error_message,
+        }
+        return render(request, "LITReviews/ticket_detail.html", context)
 
     if request.method == "POST":
         form = TicketForm(request.POST, request.FILES, instance=ticket)
         if form.is_valid():
             form.save()
-            return redirect("home")
+            context = {"form": form, "ticket": ticket}
+            return render(request, "LITReviews/ticket_detail.html", context)
     else:
         form = TicketForm(instance=ticket)
 
-    return render(
-        request, "LITReviews/modify_ticket.html", {"form": form, "ticket": ticket}
-    )
+    context = {"form": form, "ticket": ticket}
+    return render(request, "LITReviews/modify_ticket.html", context)
 
 
 def review_detail(request, review_id):
@@ -157,14 +225,28 @@ def review_detail(request, review_id):
     return render(request, "LITReviews/review_detail.html", {"review": review})
 
 
+@login_required
 def modify_review(request, review_id):
+    error = False
+    error_message = ""
     review = Review.objects.get(id=review_id)
+
+    # Check if the current user is the writer of the review or a superuser
+    if request.user != review.user and not request.user.is_superuser:
+        error = True
+        error_message = "Vous n'êtes pas autorisé(e) à modifier ce commentaire"
+        context = {
+            "review": review,
+            "error": error,
+            "error_message": error_message,
+        }
+        return render(request, "LITReviews/review_detail.html", context)
 
     if request.method == "POST":
         form = ReviewForm(request.POST, instance=review)
         if form.is_valid():
             form.save()
-            return redirect("home")
+            return redirect(request.META.get("HTTP_REFERER"))
     else:
         form = ReviewForm(instance=review)
 
@@ -172,20 +254,72 @@ def modify_review(request, review_id):
         request, "LITReviews/modify_review.html", {"form": form, "review": review}
     )
 
+
+@login_required
 def delete_review(request, review_id):
-    ticket = Review.objects.get(id=review_id)
-    if request.method == 'POST':
-        ticket.delete()
-        return redirect('post')
+    error = False
+    error_message = ""
+    try:
+        review = Review.objects.get(id=review_id)
+
+        # Check if the current user is the writer of the review or a superuser
+        if request.user != review.user and not request.user.is_superuser:
+            error = True
+            error_message = "Vous n'êtes pas autorisé(e) à supprimer ce commentaire"
+            context = {
+                "review": review,
+                "error": error,
+                "error_message": error_message,
+            }
+            return render(request, "LITReviews/review_detail.html", context)
+
+        if request.method == "POST":
+            review.delete()
+            return redirect("post")
+
+    except Review.DoesNotExist:
+        error = True
+        error_message = "Ce commentaire n'existe pas"
+        context = {
+            "error": error,
+            "error_message": error_message,
+        }
+        return render(request, "LITReviews/feed.html", context)
 
     # If the request method is not POST, render the confirmation template
-    return render(request, "LITReviews/delete_review.html", {'review': review})
+    return render(request, "LITReviews/delete_review.html", {"review": review})
 
+
+@login_required
 def delete_ticket(request, ticket_id):
-    ticket = Ticket.objects.get(id=ticket_id)
-    if request.method == 'POST':
-        ticket.delete()
-        return redirect('post')
+    error = False
+    error_message = ""
+    try:
+        ticket = Ticket.objects.get(id=ticket_id)
+
+        # Check if the current user is the writer of the ticket or a superuser
+        if request.user != ticket.user and not request.user.is_superuser:
+            error = True
+            error_message = "Vous n'êtes pas autorisé(e) à supprimer ce billet"
+            context = {
+                "ticket": ticket,
+                "error": error,
+                "error_message": error_message,
+            }
+            return render(request, "LITReviews/ticket_detail.html", context)
+
+        if request.method == "POST":
+            ticket.delete()
+            return redirect("post")
+
+    except Ticket.DoesNotExist:
+        error = True
+        error_message = "Le billet n'existe pas"
+        context = {
+            "error": error,
+            "error_message": error_message,
+        }
+        return render(request, "LITReviews/feed.html", context)
 
     # If the request method is not POST, render the confirmation template
-    return render(request, "LITReviews/delete_ticket.html", {'ticket': ticket})
+    return render(request, "LITReviews/delete_ticket.html", {"ticket": ticket})
